+++ src/nsc/MusicHeader.h
@@ -22,7 +22,7 @@ public:
 				bool	op_code;		//option で code 指定された？
 	unsigned	char	iBGM;			//BGM数
 	unsigned	char	iSE;			//SE数
-	unsigned	int		offsetPCM;
+				size_t	offsetPCM;
 				int		iExternal;		//拡張音源フラグ
 				string	Label;
 				string	title;
