+++ src/nsc/SE.cpp
@@ -16,12 +16,12 @@
 //--------------------------------------------------------------
 //	●引数
 //		MMLfile*			MML			MMLファイルのオブジェクト
-//		unsigned	int		_id			SE番号
+//					size_t	_id			SE番号
 //		const		_CHAR	_strName[]	オブジェクト名
 //	●返値
 //					無し
 //==============================================================
-SE::SE(MMLfile* MML, unsigned int _id, _CHAR _strName[]):
+SE::SE(MMLfile* MML, size_t _id, _CHAR _strName[]):
 	TrackSet(MML, _id, false, true, _strName)
 {
