+++ src/nsc/mml_Address.h
@@ -23,10 +23,10 @@ class mml_Address :
 //ƒƒ“ƒo[ŠÖ”
 public:
 	mml_Address(unsigned char _code, const _CHAR _strName[]=_T("Address"));
-	mml_Address(unsigned int _id, unsigned char _code, const _CHAR _strName[]=_T("Address"));
-	mml_Address(unsigned int _id, unsigned char _code, unsigned char _data, const _CHAR _strName[]=_T("Address"));
+	mml_Address(size_t _id, unsigned char _code, const _CHAR _strName[]=_T("Address"));
+	mml_Address(size_t _id, unsigned char _code, unsigned char _data, const _CHAR _strName[]=_T("Address"));
 	~mml_Address(void);
 
-				void	set_Address(unsigned int _addr);
-//	unsigned	int		get_Address(void);
+				void	set_Address(size_t _addr);
+//				size_t	get_Address(void);
 };
